implementations/feature3_driver_location/requirements.txt
Flask==2.2.5
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
redis==4.5.1
psycopg2-binary==2.9.5
python-dotenv==0.19.0
websocket-client==1.6.1
simple-websocket==0.5.2